+++ CHANGELOG.rst
@@ -1,11 +1,17 @@
 Changelog
 =========
 
+1.2.0 (2024-05-13)
+------------------
+* Added ``rbc.py``: A quick 'pocket calculator' for Rayleigh-Bénard Convection.
+* ``show_cli()`` will not show an end report by itself anymore. You'll have to
+  call ``report()`` afterwards yourselves.
+
 1.1.0 (2024-05-12)
 ------------------
-* Added field `Cv`: Isochoric heat capacity
-* Added field `comp`: Isothermal compressibility
+* Added field ``Cv``: Isochoric heat capacity.
+* Added field ``comp``: Isothermal compressibility.
 
 1.0.0 (2024-05-11)
 ------------------
-* Initial release
+* Initial release.